feat: simplified bot part past rendering
